requirements.txt
playwright==1.*          # Python Playwright
python-dotenv>=1.0
requests>=2.31